requirements.txt
h5py==3.16.0
numpy==1.26.4
scipy==1.15.3
trimesh==4.11.5
tqdm==4.67.3
pyyaml==6.0.3
pypdf==6.10.0
accelerate==0.28.0
mesh2sdf==1.1.0
transformers==4.39.3
opencv-python==4.9.0.80
omegaconf==2.3.0
einops==0.7.0
pyrender==0.1.45
huggingface_hub==0.36.2
scikit-image==0.24.0
lightning==2.2.5
python-box==7.3.2
timm==0.9.16
addict==2.4.0
spconv-cu118==2.3.6
